+++ R/filter_rse.R
@@ -224,3 +224,60 @@ filter_clustered_variants <- function(se, txdb,
 
     se[names(x), ]
 }
+
+
+#' Calculate confidence score for observing editing
+#'
+#' @description Calculate a confidence score based on a Bayesian inverse probability
+#' model as described by Washburn et al. Cell Reports. 2015, and implemented
+#' in the SAILOR pipeline.
+#'
+#' @param se `SummarizedExperiment::SummarizedExperiment` containing editing sites
+#' @param edit_to edited base
+#' @param edit_from non-edited base
+#' @param per_sample if TRUE, calculate confidence per sample, otherwise edited
+#' and non-edited counts will be summed across all samples.
+#' @param exp_fraction Numeric, confidence margin parameter for
+#'
+#' @examples
+#' data(rse_adar_ifn)
+#' calc_confidence(rse_adar_ifn)
+#' calc_confidence(rse_adar_ifn, per_sample = TRUE)
+#'
+#' @return `SummarizedExperiment::SummarizedExperiment` with either a new assay
+#' or rowData column named "confidence" depending on whether confidence is
+#'  calculated `per_sample`.
+#'
+#' @references
+#' Washburn MC, Kakaradov B, Sundararaman B, Wheeler E, Hoon S, Yeo GW, Hundley HA. The dsRBP and inactive editor ADR-1 utilizes dsRNA binding to regulate A-to-I RNA editing across the C. elegans transcriptome. Cell Rep. 2014 Feb 27;6(4):599-607. doi: 10.1016/j.celrep.2014.01.011. Epub 2014 Feb 6. PMID: 24508457; PMCID: PMC3959997.
+#'
+#' SAILOR pipeline: https://github.com/YeoLab/sailor
+#' @importFrom stats pbeta
+#' @export
+calc_confidence <- function(se,
+                            edit_to = "G",
+                            edit_from = "A",
+                            per_sample = FALSE,
+                            exp_fraction = 0.01){
+  if(length(exp_fraction) != 1) {
+    cli::cli_abort("exp_fraction must be numeric(1)")
+  }
+  edit_to <- paste0("n", edit_to)
+  edit_from <- paste0("n", edit_from)
+  alt <- assay(se, edit_to)
+  ref <- assay(se, edit_from)
+  if(per_sample){
+    nc <- ncol(se)
+    res <- vapply(seq_len(nc), function(i){
+      1 - pbeta(exp_fraction, alt[, i], ref[, i])
+    },  FUN.VALUE = numeric(nrow(se)))
+    colnames(res) <- colnames(se)
+    assays(se)$confidence <- res
+  } else {
+    alt <- rowSums(alt)
+    ref <- rowSums(ref)
+    res <- 1 - pbeta(exp_fraction, alt, ref)
+    rowData(se)$confidence <- res
+  }
+  se
+}